adding updates to DOA and API

+++ 07_Data_Access_and_DAO/exercise/src/main/java/com/techelevator/projects/model/Department.java
@@ -2,12 +2,12 @@ package com.techelevator.projects.model;
 
 public class Department {
 
-	private int id;
+	private Long id;
 	private String name;
 
 	public Department() {}
 
-	public Department(int id, String name) {
+	public Department(Long id, String name) {
 		this.id = id;
 		this.name = name;
 	}
@@ -20,11 +20,11 @@ public class Department {
 		this.name = name;
 	}
 
-	public int getId() {
+	public Long getId() {
 		return id;
 	}
 	
-	public void setId(int id) {
+	public void setId(Long id) {
 		this.id = id;
 	}
